+++ summarization.py
@@ -44,7 +44,7 @@ examples = [
     AIMessage(
         dedent(
             """
-            트럼프 대통령은 관세 정책에 변화가 없음을 재확인하며, 반도체 및 전자제품에 대한 기존 20% 펜타닐 관세를 유지하고 국가 안보 조사 시반도체와 전자제품 공급망을 검토할 것임을 발표하였습니다. 
+            트럼프 대통령은 관세 정책에 변화가 없음을 재확인하며, 반도체 및 전자제품에 대한 기존 20% 펜타닐 관세를 유지하고 국가 안보 조사 시 반도체와 전자제품 공급망을 검토할 것임을 발표하였습니다. 
             이는 중국에 부과된 상호관세의 일부 면제에도 불구한 조치입니다. 
             트럼프 대통령은 모든 국가, 특히 미국을 최악으로 대우하는 중국에 대해 관세 정책을 계속 추진할 것임을 강조했습니다.
             """
